+++ internal/core/ports/irecipe.go
@@ -8,10 +8,10 @@ import (
 
 // IRecipeRepository define las operaciones para el repositorio de recetas
 type IRecipeRepository interface {
-	GetRecipesByAge(ctx context.Context, age float32) ([]*domain.Recipe, error)
+	GetRecipesByAge(ctx context.Context, age float64) ([]*domain.Recipe, error)
 }
 
 // IRecipeService define las operaciones del servicio para recetas
 type IRecipeService interface {
-	ListRecipesByAge(ctx context.Context, age float32) ([]*domain.Recipe, error)
+	ListRecipesByAge(ctx context.Context, age float64) ([]*domain.Recipe, error)
 }
